refactor: apply review suggestions
* corrected error message to reflect xsp magic
* moved utils into msixvc::models::common
* moved common traits into msixvc::models::common
* removed msixvc::utils::{xsp, xvd} nesting
