+++ msixvc/src/lib.rs
@@ -4,4 +4,5 @@ pub mod math;
 pub mod models;
 pub mod streaming;
 pub mod streaming_ntfs;
-pub mod utils;
+pub mod xsp;
+pub mod xvd;
