+++ msixvc/src/models.rs
@@ -1,2 +1,3 @@
+pub mod common;
 pub mod xsp;
 pub mod xvd;
