--- msixvc/src/utils.rs
@@ -1,7 +1,0 @@
-pub mod common;
-pub mod xsp;
-pub mod xvd;
-
-pub use common::*;
-pub use xsp::*;
-pub use xvd::*;
