+++ xodus-cli/src/commands/extract.rs
@@ -1,4 +1,4 @@
-use msixvc::utils::{XvdFile, unpack_file};
+use msixvc::xvd::{XvdFile, unpack_file};
 use xodus::tokens::TokenManager;
 
 use crate::license::get_license;
